model-server/requirements.txt
Flask==1.0.2
Flask-Cors==3.0.6
numpy==1.15.2
opencv-python==3.4.3.18
tensorflow-gpu==1.11.0
